Change entity id to Long. Create BaseEntity, NamedEntity

+++ src/main/java/ua/gaponov/school/feature/academicterm/AcademicTermRepo.java
@@ -2,6 +2,6 @@ package ua.gaponov.school.feature.academicterm;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AcademicTermRepo extends JpaRepository<AcademicTerm, Integer> {
+public interface AcademicTermRepo extends JpaRepository<AcademicTerm, Long> {
 
 }
