+++ src/main/java/ua/gaponov/school/feature/academicyear/AcademicYearRepoBasic.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface AcademicYearRepoBasic extends JpaRepository<AcademicYear, Integer> {
+public interface AcademicYearRepoBasic extends JpaRepository<AcademicYear, Long> {
   @Query(nativeQuery = true, value = "SELECT * FROM academic_year y WHERE y.school_id = :id")
-  List<AcademicYear> findBySchoolId(int id);
+  List<AcademicYear> findBySchoolId(long id);
 }
