+++ src/main/java/ua/gaponov/school/feature/courses/CourseRepoBasic.java
@@ -2,6 +2,6 @@ package ua.gaponov.school.feature.courses;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseRepoBasic  extends JpaRepository<Course, Integer> {
+public interface CourseRepoBasic  extends JpaRepository<Course, Long> {
 
 }
