+++ src/main/java/ua/gaponov/school/feature/grades/GradeRepo.java
@@ -2,6 +2,6 @@ package ua.gaponov.school.feature.grades;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GradeRepo extends JpaRepository<Grade, Integer> {
+public interface GradeRepo extends JpaRepository<Grade, Long> {
 
 }
