+++ src/main/java/ua/gaponov/school/feature/school/SchoolRepoBasic.java
@@ -2,6 +2,6 @@ package ua.gaponov.school.feature.school;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SchoolRepoBasic extends JpaRepository<School, Integer> {
+public interface SchoolRepoBasic extends JpaRepository<School, Long> {
 
 }
