+++ src/main/java/ua/gaponov/school/feature/schoolclass/SchoolClassRepoCustom.java
@@ -5,6 +5,6 @@ import java.util.List;
 public interface SchoolClassRepoCustom {
 
   List<SchoolClass> findAllByKeywords(String keywords);
-  List<SchoolClass> findAllBySchoolId(int id);
-  List<SchoolClass> findAllBySchoolIdAndKeywords(int id, String keywords);
+  List<SchoolClass> findAllBySchoolId(long id);
+  List<SchoolClass> findAllBySchoolIdAndKeywords(long id, String keywords);
 }
