+++ src/main/java/ua/gaponov/school/feature/students/StudentRepoBasic.java
@@ -2,6 +2,6 @@ package ua.gaponov.school.feature.students;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepoBasic extends JpaRepository<Student, Integer> {
+public interface StudentRepoBasic extends JpaRepository<Student, Long> {
 
 }
